apps sends info to Callback server
When an app sends a request to IMPACT platform, it receives an immediate
response with requestId in its body. Apps sends this info together with
its URL to callback server so that when IMPACT sends th actual data to
callback server, callback server knows to which app the info should be redirected.