config setting chnaged
Now at start-up, RE checks if a config file exists (at config) folder or not. If not, it creates it by copying from config-template.json file. Some config settings are comming form environment vaiables (in case of Openshift deployment env vaiables are present), but if not present (in case of local deployment), RE initializes env variables in the code.