This commit was a try to make update of app working
While working on the update app, I understand that it is not possible to
completely clean a nodejs module if module is running in main process
and not in a separate process. As user applications are runnings as
nodejs modules in the main process, the whole concept so far was based a
wrong decition. for further info lokk at
https://stackoverflow.com/questions/47392613/how-to-completely-remove-a-nodejs-module-with-all-its-functionalities-even-inclu