fix a bug: a killed application should not activate update app info on the DB
when an app is killed by killing it pid, it activates the 'exit' listener with 2 code.
When an app crashes, it activates the 'exit' listener with 99 code.
Only on the case of crashing the app info on the DB should be updated,
Since app deletion, the app info is already deleted from DB.